pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "academic-slides-agent"
version = "0.1.0"
description = "Multi-agent pipeline for generating high-fidelity 4K academic presentations from research papers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "keyring>=25.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "anthropic>=0.30",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
slides = "academic_slides_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
